+++ Backend/src/ConsertaPraMim.Domain/Repositories/IAdminAuditLogRepository.cs
@@ -1,0 +1,8 @@
+using ConsertaPraMim.Domain.Entities;
+
+namespace ConsertaPraMim.Domain.Repositories;
+
+public interface IAdminAuditLogRepository
+{
+    Task AddAsync(AdminAuditLog auditLog);
+}